requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-cors==0.7.0
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
beartype==0.15.0
bleach==6.1.0
blinker==1.8.2
bokeh==3.5.2
boto3==1.35.20
botocore==1.35.20
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cohere==5.9.4
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.14
diskcache==5.6.3
distro==1.9.0
docarray==0.40.0
executing==2.1.0
fastapi==0.115.0
fastavro==1.9.7
ffmpy==0.4.0
filelock==3.16.0
fonttools==4.53.1
frozenlist==1.4.1
fs==2.4.16
fsspec==2024.9.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-pubsub==2.23.1
google-crc32c==1.6.0
google-generativeai==0.8.1
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gradio==4.44.0
gradio_client==1.3.0
greenlet==3.1.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.62.3
h11==0.14.0
h3==3.7.7
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.24.7
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_bokeh==4.0.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-mistralai==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.121
linkify-it-py==2.0.3
litellm==1.46.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.8.1
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
openai==1.45.1
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
panel==1.5.0
param==2.1.1
parameterized==0.9.0
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pinecone==5.3.0
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.3
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.7.4
pydantic-settings==2.5.2
pydantic_core==2.18.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-sat==1.8.dev13
pytz==2024.2
pyviz_comms==3.0.3
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
ruff==0.6.6
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
sentence-transformers==3.1.0
setuptools==75.1.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlglot==10.6.1
stack-data==0.6.3
starlette==0.38.5
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
triton==3.0.0
typer==0.12.5
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
watchdog==4.0.2
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
widgetsnbextension==4.0.13
wrapt==1.16.0
xyzservices==2024.9.0
yarl==1.11.1
zipp==3.20.2
